Allow multiple react/react-native versions (#26)
Hey guys, awesome work on this library. I couldn't find anything in the
documentation as to why we had to pin to react-native 0.81.0
specifically (as opposed to react-native 0.81.4, for instance) and same
for React. On my team, we're on react native 0.81.4 and React 19.2.0 and
would still like to use this library, and I imagine others might be in
the same boat.

If this is off base feel free to let me know! 
Again, really appreciate the work that's gone into this package. 

**Contributing to Twilio**

> All third-party contributors acknowledge that any contributions they
provide will be made under the same open-source license that the
open-source project is provided under.

- [x] I acknowledge that all my contributions will be made under the
project's license.

---------

Co-authored-by: zhanim-tdl <zhani.muceku@testdevlab.com>